refactor: inherit all from PgStore
