+++ src/api/routes/contract.ts
@@ -3,7 +3,7 @@ import { asyncHandler } from '../async-handler';
 import { parseLimitQuery, parsePagingQueryInput } from '../pagination';
 import { parseDbEvent } from '../controllers/db-controller';
 import { parseTraitAbi } from '../query-helpers';
-import { PgReplicaStore } from '../../datastore/pg-replica-store';
+import { PgStore } from '../../datastore/pg-store';
 
 const MAX_EVENTS_PER_REQUEST = 50;
 const parseContractEventsQueryLimit = parseLimitQuery({
@@ -11,7 +11,7 @@ const parseContractEventsQueryLimit = parseLimitQuery({
   errorMsg: '`limit` must be equal to or less than ' + MAX_EVENTS_PER_REQUEST,
 });
 
-export function createContractRouter(db: PgReplicaStore): express.Router {
+export function createContractRouter(db: PgStore): express.Router {
   const router = express.Router();
 
   router.get(
