+++ src/api/routes/status.ts
@@ -3,9 +3,9 @@ import * as fs from 'fs';
 import { ServerStatusResponse } from '@stacks/stacks-blockchain-api-types';
 import { logger } from '../../helpers';
 import { getETagCacheHandler, setETagCacheHeaders } from '../controllers/cache-controller';
-import { PgReplicaStore } from '../../datastore/pg-replica-store';
+import { PgStore } from '../../datastore/pg-store';
 
-export function createStatusRouter(db: PgReplicaStore): express.Router {
+export function createStatusRouter(db: PgStore): express.Router {
   const router = express.Router();
   const cacheHandler = getETagCacheHandler(db);
   const statusHandler = async (_: Request, res: any) => {
