+++ src/datastore/connection.ts
@@ -1,0 +1,202 @@
+import { Client, ClientConfig, Pool, PoolClient, PoolConfig } from 'pg';
+import { logError, logger, parseArgBoolean, parsePort, stopwatch, timeout } from '../helpers';
+
+export type PgClientConfig = ClientConfig & { schema?: string };
+export type PgPoolConfig = PoolConfig & { schema?: string };
+
+/**
+ * The postgres server being used for a particular connection, transaction or query.
+ * The API will automatically choose between `default` (or read replica) and `primary`.
+ * See `.env` for more information.
+ */
+export enum PgServer {
+  default,
+  primary,
+}
+
+export async function connectPgPool({
+  usageName,
+  pgServer,
+}: {
+  usageName: string;
+  pgServer: PgServer;
+}): Promise<Pool> {
+  const initTimer = stopwatch();
+  let connectionError: Error | undefined;
+  let connectionOkay = false;
+  let lastElapsedLog = 0;
+  do {
+    const clientConfig = getPgClientConfig({
+      usageName: `${usageName};init-connection-poll`,
+      primary: pgServer === PgServer.primary,
+    });
+    const client = new Client(clientConfig);
+    try {
+      await client.connect();
+      connectionOkay = true;
+      break;
+    } catch (error: any) {
+      const pgConnectionError = isPgConnectionError(error);
+      if (!pgConnectionError) {
+        logError('Cannot connect to pg', error);
+        throw error;
+      }
+      const timeElapsed = initTimer.getElapsed();
+      if (timeElapsed - lastElapsedLog > 2000) {
+        lastElapsedLog = timeElapsed;
+        logError('Pg connection failed, retrying..');
+      }
+      connectionError = error;
+      await timeout(100);
+    } finally {
+      client.end(() => {});
+    }
+  } while (initTimer.getElapsed() < Number.MAX_SAFE_INTEGER);
+  if (!connectionOkay) {
+    connectionError = connectionError ?? new Error('Error connecting to database');
+    throw connectionError;
+  }
+  const poolConfig: PoolConfig = getPgClientConfig({
+    usageName: `${usageName};datastore-crud`,
+    getPoolConfig: true,
+    primary: pgServer === PgServer.primary,
+  });
+  const pool = new Pool(poolConfig);
+  pool.on('error', error => {
+    logger.error(`Postgres connection pool error: ${error.message}`, error);
+  });
+  return pool;
+}
+
+/**
+ * @typeParam TGetPoolConfig - If specified as true, returns a PoolConfig object where max connections are configured. Otherwise, returns a regular ClientConfig.
+ */
+export function getPgClientConfig<TGetPoolConfig extends boolean = false>({
+  usageName,
+  primary = false,
+  getPoolConfig,
+}: {
+  usageName: string;
+  primary?: boolean;
+  getPoolConfig?: TGetPoolConfig;
+}): TGetPoolConfig extends true ? PgPoolConfig : PgClientConfig {
+  // Retrieve a postgres ENV value depending on the target database server (read-replica/default or primary).
+  // We will fall back to read-replica values if a primary value was not given.
+  // See the `.env` file for more information on these options.
+  const pgEnvValue = (name: string): string | undefined =>
+    primary
+      ? process.env[`PG_PRIMARY_${name}`] ?? process.env[`PG_${name}`]
+      : process.env[`PG_${name}`];
+  const pgEnvVars = {
+    database: pgEnvValue('DATABASE'),
+    user: pgEnvValue('USER'),
+    password: pgEnvValue('PASSWORD'),
+    host: pgEnvValue('HOST'),
+    port: pgEnvValue('PORT'),
+    ssl: pgEnvValue('SSL'),
+    schema: pgEnvValue('SCHEMA'),
+    applicationName: pgEnvValue('APPLICATION_NAME'),
+  };
+  const defaultAppName = 'stacks-blockchain-api';
+  const pgConnectionUri = pgEnvValue('CONNECTION_URI');
+  const pgConfigEnvVar = Object.entries(pgEnvVars).find(([, v]) => typeof v === 'string')?.[0];
+  if (pgConfigEnvVar && pgConnectionUri) {
+    throw new Error(
+      `Both PG_CONNECTION_URI and ${pgConfigEnvVar} environmental variables are defined. PG_CONNECTION_URI must be defined without others or omitted.`
+    );
+  }
+  let clientConfig: PgClientConfig;
+  if (pgConnectionUri) {
+    const uri = new URL(pgConnectionUri);
+    const searchParams = Object.fromEntries(
+      [...uri.searchParams.entries()].map(([k, v]) => [k.toLowerCase(), v])
+    );
+    // Not really standardized
+    const schema: string | undefined =
+      searchParams['currentschema'] ??
+      searchParams['current_schema'] ??
+      searchParams['searchpath'] ??
+      searchParams['search_path'] ??
+      searchParams['schema'];
+    const appName = `${uri.searchParams.get('application_name') ?? defaultAppName}:${usageName}`;
+    uri.searchParams.set('application_name', appName);
+    clientConfig = {
+      connectionString: uri.toString(),
+      schema,
+    };
+  } else {
+    const appName = `${pgEnvVars.applicationName ?? defaultAppName}:${usageName}`;
+    clientConfig = {
+      database: pgEnvVars.database,
+      user: pgEnvVars.user,
+      password: pgEnvVars.password,
+      host: pgEnvVars.host,
+      port: parsePort(pgEnvVars.port),
+      ssl: parseArgBoolean(pgEnvVars.ssl),
+      schema: pgEnvVars.schema,
+      application_name: appName,
+    };
+  }
+  if (getPoolConfig) {
+    const poolConfig: PgPoolConfig = { ...clientConfig };
+    const pgConnectionPoolMaxEnv = process.env['PG_CONNECTION_POOL_MAX'];
+    if (pgConnectionPoolMaxEnv) {
+      poolConfig.max = Number.parseInt(pgConnectionPoolMaxEnv);
+    }
+    return poolConfig;
+  } else {
+    return clientConfig;
+  }
+}
+
+/**
+ * Creates a postgres pool client connection. If the connection fails due to a transient error, it is retried until successful.
+ * You'd expect that the pg lib to handle this, but it doesn't, see https://github.com/brianc/node-postgres/issues/1789
+ */
+export async function connectWithRetry(pool: Pool): Promise<PoolClient> {
+  for (let retryAttempts = 1; ; retryAttempts++) {
+    try {
+      const client = await pool.connect();
+      return client;
+    } catch (error: any) {
+      // Check for transient errors, and retry after 1 second
+      const pgConnectionError = isPgConnectionError(error);
+      if (pgConnectionError) {
+        logger.warn(`${pgConnectionError}, will retry, attempt #${retryAttempts}`);
+        await timeout(1000);
+      } else {
+        throw error;
+      }
+    }
+  }
+}
+
+/**
+ * Checks if a given error from the pg lib is a connection error (i.e. the query is retryable).
+ * If true then returns a normalized error message, otherwise returns false.
+ */
+function isPgConnectionError(error: any): string | false {
+  if (error.code === 'ECONNREFUSED') {
+    return 'Postgres connection ECONNREFUSED';
+  } else if (error.code === 'ETIMEDOUT') {
+    return 'Postgres connection ETIMEDOUT';
+  } else if (error.code === 'ENOTFOUND') {
+    return 'Postgres connection ENOTFOUND';
+  } else if (error.message) {
+    const msg = (error as Error).message.toLowerCase();
+    if (msg.includes('database system is starting up')) {
+      return 'Postgres connection failed while database system is starting up';
+    } else if (msg.includes('database system is shutting down')) {
+      return 'Postgres connection failed while database system is shutting down';
+    } else if (msg.includes('connection terminated unexpectedly')) {
+      return 'Postgres connection terminated unexpectedly';
+    } else if (msg.includes('connection terminated')) {
+      return 'Postgres connection terminated';
+    } else if (msg.includes('connection error')) {
+      return 'Postgres client has encountered a connection error and is not queryable';
+    } else if (msg.includes('terminating connection due to unexpected postmaster exit')) {
+      return 'Postgres connection terminating due to unexpected postmaster exit';
+    }
+  }
+  return false;
+}
