+++ src/datastore/helpers.ts
@@ -30,9 +30,6 @@ import {
   DbTxTypeId,
 } from './common';
 
-const MIGRATIONS_TABLE = 'pgmigrations';
-const MIGRATIONS_DIR = path.join(APP_DIR, 'migrations');
-
 export interface BlockQueryResult {
   block_hash: Buffer;
   index_block_hash: Buffer;
@@ -330,216 +327,6 @@ export interface SmartContractLogEventResult {
 
 export interface RawTxQueryResult {
   raw_tx: Buffer;
-}
-
-type PgClientConfig = ClientConfig & { schema?: string };
-type PgPoolConfig = PoolConfig & { schema?: string };
-
-/**
- * The postgres server being used for a particular connection, transaction or query.
- * The API will automatically choose between `default` (or read replica) and `primary`.
- * See `.env` for more information.
- */
-enum PgServer {
-  default,
-  primary,
-}
-
-/**
- * @typeParam TGetPoolConfig - If specified as true, returns a PoolConfig object where max connections are configured. Otherwise, returns a regular ClientConfig.
- */
-export function getPgClientConfig<TGetPoolConfig extends boolean = false>({
-  usageName,
-  primary = false,
-  getPoolConfig,
-}: {
-  usageName: string;
-  primary?: boolean;
-  getPoolConfig?: TGetPoolConfig;
-}): TGetPoolConfig extends true ? PgPoolConfig : PgClientConfig {
-  // Retrieve a postgres ENV value depending on the target database server (read-replica/default or primary).
-  // We will fall back to read-replica values if a primary value was not given.
-  // See the `.env` file for more information on these options.
-  const pgEnvValue = (name: string): string | undefined =>
-    primary
-      ? process.env[`PG_PRIMARY_${name}`] ?? process.env[`PG_${name}`]
-      : process.env[`PG_${name}`];
-  const pgEnvVars = {
-    database: pgEnvValue('DATABASE'),
-    user: pgEnvValue('USER'),
-    password: pgEnvValue('PASSWORD'),
-    host: pgEnvValue('HOST'),
-    port: pgEnvValue('PORT'),
-    ssl: pgEnvValue('SSL'),
-    schema: pgEnvValue('SCHEMA'),
-    applicationName: pgEnvValue('APPLICATION_NAME'),
-  };
-  const defaultAppName = 'stacks-blockchain-api';
-  const pgConnectionUri = pgEnvValue('CONNECTION_URI');
-  const pgConfigEnvVar = Object.entries(pgEnvVars).find(([, v]) => typeof v === 'string')?.[0];
-  if (pgConfigEnvVar && pgConnectionUri) {
-    throw new Error(
-      `Both PG_CONNECTION_URI and ${pgConfigEnvVar} environmental variables are defined. PG_CONNECTION_URI must be defined without others or omitted.`
-    );
-  }
-  let clientConfig: PgClientConfig;
-  if (pgConnectionUri) {
-    const uri = new URL(pgConnectionUri);
-    const searchParams = Object.fromEntries(
-      [...uri.searchParams.entries()].map(([k, v]) => [k.toLowerCase(), v])
-    );
-    // Not really standardized
-    const schema: string | undefined =
-      searchParams['currentschema'] ??
-      searchParams['current_schema'] ??
-      searchParams['searchpath'] ??
-      searchParams['search_path'] ??
-      searchParams['schema'];
-    const appName = `${uri.searchParams.get('application_name') ?? defaultAppName}:${usageName}`;
-    uri.searchParams.set('application_name', appName);
-    clientConfig = {
-      connectionString: uri.toString(),
-      schema,
-    };
-  } else {
-    const appName = `${pgEnvVars.applicationName ?? defaultAppName}:${usageName}`;
-    clientConfig = {
-      database: pgEnvVars.database,
-      user: pgEnvVars.user,
-      password: pgEnvVars.password,
-      host: pgEnvVars.host,
-      port: parsePort(pgEnvVars.port),
-      ssl: parseArgBoolean(pgEnvVars.ssl),
-      schema: pgEnvVars.schema,
-      application_name: appName,
-    };
-  }
-  if (getPoolConfig) {
-    const poolConfig: PgPoolConfig = { ...clientConfig };
-    const pgConnectionPoolMaxEnv = process.env['PG_CONNECTION_POOL_MAX'];
-    if (pgConnectionPoolMaxEnv) {
-      poolConfig.max = Number.parseInt(pgConnectionPoolMaxEnv);
-    }
-    return poolConfig;
-  } else {
-    return clientConfig;
-  }
-}
-
-export async function runMigrations(
-  clientConfig: PgClientConfig = getPgClientConfig({ usageName: 'schema-migrations' }),
-  direction: 'up' | 'down' = 'up',
-  opts?: {
-    // Bypass the NODE_ENV check when performing a "down" migration which irreversibly drops data.
-    dangerousAllowDataLoss?: boolean;
-  }
-): Promise<void> {
-  if (!opts?.dangerousAllowDataLoss && direction !== 'up' && !isTestEnv && !isDevEnv) {
-    throw new Error(
-      'Whoa there! This is a testing function that will drop all data from PG. ' +
-        'Set NODE_ENV to "test" or "development" to enable migration testing.'
-    );
-  }
-  const client = new Client(clientConfig);
-  try {
-    await client.connect();
-    const runnerOpts: RunnerOption = {
-      dbClient: client,
-      ignorePattern: '.*map',
-      dir: MIGRATIONS_DIR,
-      direction: direction,
-      migrationsTable: MIGRATIONS_TABLE,
-      count: Infinity,
-      logger: {
-        info: msg => {},
-        warn: msg => logger.warn(msg),
-        error: msg => logger.error(msg),
-      },
-    };
-    if (clientConfig.schema) {
-      runnerOpts.schema = clientConfig.schema;
-    }
-    await PgMigrate(runnerOpts);
-  } catch (error) {
-    logError(`Error running pg-migrate`, error);
-    throw error;
-  } finally {
-    await client.end();
-  }
-}
-
-export async function cycleMigrations(opts?: {
-  // Bypass the NODE_ENV check when performing a "down" migration which irreversibly drops data.
-  dangerousAllowDataLoss?: boolean;
-}): Promise<void> {
-  const clientConfig = getPgClientConfig({ usageName: 'cycle-migrations' });
-
-  await runMigrations(clientConfig, 'down', opts);
-  await runMigrations(clientConfig, 'up', opts);
-}
-
-export async function dangerousDropAllTables(opts?: {
-  acknowledgePotentialCatastrophicConsequences?: 'yes';
-}) {
-  if (opts?.acknowledgePotentialCatastrophicConsequences !== 'yes') {
-    throw new Error('Dangerous usage error.');
-  }
-  const clientConfig = getPgClientConfig({ usageName: 'dangerous-drop-all-tables' });
-  const client = new Client(clientConfig);
-  try {
-    await client.connect();
-    await client.query('BEGIN');
-    const getTablesQuery = await client.query<{ table_name: string }>(
-      `
-      SELECT table_name
-      FROM information_schema.tables
-      WHERE table_schema = $1
-      AND table_catalog = $2
-      AND table_type = 'BASE TABLE'
-      `,
-      [clientConfig.schema, clientConfig.database]
-    );
-    const tables = getTablesQuery.rows.map(r => r.table_name);
-    for (const table of tables) {
-      await client.query(`DROP TABLE IF EXISTS ${table} CASCADE`);
-    }
-    await client.query('COMMIT');
-  } catch (error) {
-    await client.query('ROLLBACK');
-    throw error;
-  } finally {
-    await client.end();
-  }
-}
-
-/**
- * Checks if a given error from the pg lib is a connection error (i.e. the query is retryable).
- * If true then returns a normalized error message, otherwise returns false.
- */
-export function isPgConnectionError(error: any): string | false {
-  if (error.code === 'ECONNREFUSED') {
-    return 'Postgres connection ECONNREFUSED';
-  } else if (error.code === 'ETIMEDOUT') {
-    return 'Postgres connection ETIMEDOUT';
-  } else if (error.code === 'ENOTFOUND') {
-    return 'Postgres connection ENOTFOUND';
-  } else if (error.message) {
-    const msg = (error as Error).message.toLowerCase();
-    if (msg.includes('database system is starting up')) {
-      return 'Postgres connection failed while database system is starting up';
-    } else if (msg.includes('database system is shutting down')) {
-      return 'Postgres connection failed while database system is shutting down';
-    } else if (msg.includes('connection terminated unexpectedly')) {
-      return 'Postgres connection terminated unexpectedly';
-    } else if (msg.includes('connection terminated')) {
-      return 'Postgres connection terminated';
-    } else if (msg.includes('connection error')) {
-      return 'Postgres client has encountered a connection error and is not queryable';
-    } else if (msg.includes('terminating connection due to unexpected postmaster exit')) {
-      return 'Postgres connection terminating due to unexpected postmaster exit';
-    }
-  }
-  return false;
 }
 
 /**
