+++ src/datastore/migrations.ts
@@ -1,0 +1,94 @@
+import * as path from 'path';
+import PgMigrate, { RunnerOption } from 'node-pg-migrate';
+import { Client } from 'pg';
+import { APP_DIR, isDevEnv, isTestEnv, logError, logger } from '../helpers';
+import { getPgClientConfig, PgClientConfig } from './connection';
+
+const MIGRATIONS_TABLE = 'pgmigrations';
+const MIGRATIONS_DIR = path.join(APP_DIR, 'migrations');
+
+export async function runMigrations(
+  clientConfig: PgClientConfig = getPgClientConfig({ usageName: 'schema-migrations' }),
+  direction: 'up' | 'down' = 'up',
+  opts?: {
+    // Bypass the NODE_ENV check when performing a "down" migration which irreversibly drops data.
+    dangerousAllowDataLoss?: boolean;
+  }
+): Promise<void> {
+  if (!opts?.dangerousAllowDataLoss && direction !== 'up' && !isTestEnv && !isDevEnv) {
+    throw new Error(
+      'Whoa there! This is a testing function that will drop all data from PG. ' +
+        'Set NODE_ENV to "test" or "development" to enable migration testing.'
+    );
+  }
+  const client = new Client(clientConfig);
+  try {
+    await client.connect();
+    const runnerOpts: RunnerOption = {
+      dbClient: client,
+      ignorePattern: '.*map',
+      dir: MIGRATIONS_DIR,
+      direction: direction,
+      migrationsTable: MIGRATIONS_TABLE,
+      count: Infinity,
+      logger: {
+        info: msg => {},
+        warn: msg => logger.warn(msg),
+        error: msg => logger.error(msg),
+      },
+    };
+    if (clientConfig.schema) {
+      runnerOpts.schema = clientConfig.schema;
+    }
+    await PgMigrate(runnerOpts);
+  } catch (error) {
+    logError(`Error running pg-migrate`, error);
+    throw error;
+  } finally {
+    await client.end();
+  }
+}
+
+export async function cycleMigrations(opts?: {
+  // Bypass the NODE_ENV check when performing a "down" migration which irreversibly drops data.
+  dangerousAllowDataLoss?: boolean;
+}): Promise<void> {
+  const clientConfig = getPgClientConfig({ usageName: 'cycle-migrations' });
+
+  await runMigrations(clientConfig, 'down', opts);
+  await runMigrations(clientConfig, 'up', opts);
+}
+
+export async function dangerousDropAllTables(opts?: {
+  acknowledgePotentialCatastrophicConsequences?: 'yes';
+}) {
+  if (opts?.acknowledgePotentialCatastrophicConsequences !== 'yes') {
+    throw new Error('Dangerous usage error.');
+  }
+  const clientConfig = getPgClientConfig({ usageName: 'dangerous-drop-all-tables' });
+  const client = new Client(clientConfig);
+  try {
+    await client.connect();
+    await client.query('BEGIN');
+    const getTablesQuery = await client.query<{ table_name: string }>(
+      `
+      SELECT table_name
+      FROM information_schema.tables
+      WHERE table_schema = $1
+      AND table_catalog = $2
+      AND table_type = 'BASE TABLE'
+      `,
+      [clientConfig.schema, clientConfig.database]
+    );
+    const tables = getTablesQuery.rows.map(r => r.table_name);
+    for (const table of tables) {
+      await client.query(`DROP TABLE IF EXISTS ${table} CASCADE`);
+    }
+    await client.query('COMMIT');
+  } catch (error) {
+    await client.query('ROLLBACK');
+    throw error;
+  } finally {
+    await client.end();
+  }
+}
